#10334: Migrate recip_bw to TTNN with overloading
