+++ tt_eager/tt_dnn/op_library/complex/complex_ops.cpp
@@ -297,26 +297,6 @@ std::vector<ComplexTensor> complex_div_bw(const ComplexTensor& grad, const Compl
     return grad_tensor;
 }
 
-// complex reciprocal
-// self: -grad * (result * result).conj()
-std::vector<ComplexTensor> complex_recip_bw(const ComplexTensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config) {
-    std::vector<ComplexTensor> grad_tensor;
-    Tensor condition_nan = ttnn::logical_and(ttnn::eqz(input.real(),output_mem_config), ttnn::eqz(input.imag(),output_mem_config), std::nullopt, output_mem_config);
-    ComplexTensor neg_grad = ComplexTensor({ttnn::neg(grad.real(),output_mem_config), ttnn::neg(grad.imag(),output_mem_config)});
-    ComplexTensor inp_recip = complex_recip(input, output_mem_config);
-    ComplexTensor grad_inp = complex_mul(neg_grad, conj(complex_mul(inp_recip, inp_recip, output_mem_config), output_mem_config), output_mem_config) ;
-    neg_grad.deallocate();
-    inp_recip.deallocate();
-    Tensor grad_inp_r = where(condition_nan, full_like(input.real(), std::nanf(""), output_mem_config), grad_inp.real(), output_mem_config);
-    Tensor grad_inp_i = where(condition_nan, full_like(input.imag(), std::nanf(""), output_mem_config), grad_inp.imag(), output_mem_config);
-    condition_nan.deallocate();
-    grad_inp = ComplexTensor({ grad_inp_r, grad_inp_i});
-    grad_inp_r.deallocate();
-    grad_inp_i.deallocate();
-    grad_tensor.emplace_back(grad_inp);
-    return grad_tensor;
-}
-
 }//namespace tt_metal
 
 }//namespace tt
