+++ ttnn/cpp/ttnn/operations/eltwise/complex_unary_backward/device/complex_unary_backward_op.hpp
@@ -20,11 +20,13 @@ enum class ComplexUnaryBackwardOpType {
     ANGLE_BW,
     CONJ_BW,
     COMPLEX_ABS_BW,
+    COMPLEX_RECIP_BW,
 };
 
 //OpHandler_complex : get_function_complex
 std::vector<ComplexTensor> _polar_bw(const ComplexTensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config);
 std::vector<ComplexTensor> _conj_bw(const ComplexTensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config);
+std::vector<ComplexTensor> _complex_recip_bw(const ComplexTensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config);
 
 //OpHandler_tensor_complex : get_function_tensor_complex
 std::vector<ComplexTensor> _imag_bw(const Tensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config);
@@ -49,6 +51,13 @@ template <>
 struct OpHandler_complex<ComplexUnaryBackwardOpType::CONJ_BW> {
     static std::vector<ComplexTensor> handle( const ComplexTensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config ) {
         return _conj_bw(grad, input, output_mem_config);
+    }
+};
+
+template <>
+struct OpHandler_complex<ComplexUnaryBackwardOpType::COMPLEX_RECIP_BW> {
+    static std::vector<ComplexTensor> handle( const ComplexTensor& grad, const ComplexTensor& input, const MemoryConfig& output_mem_config ) {
+        return _complex_recip_bw(grad, input, output_mem_config);
     }
 };
