+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/unary_backward_pybind.hpp
@@ -572,6 +572,19 @@ Example:
 
         ttnn::pybind_overload_t{
             [operation](const unary_backward_operation_t& self,
+               const ComplexTensor& grad_tensor,
+               const ComplexTensor& input_tensor_a,
+               const MemoryConfig& memory_config) -> std::vector<ComplexTensor> {
+                using ComplexUnaryBackwardOp = ttnn::operations::complex_unary_backward::ExecuteComplexUnaryBackward<complex_unary_backward::ComplexUnaryBackwardOpType::COMPLEX_RECIP_BW>;
+                return ComplexUnaryBackwardOp::execute_on_main_thread(grad_tensor, input_tensor_a, memory_config);
+            },
+            py::arg("grad_tensor"),
+            py::arg("input_tensor_a"),
+            py::kw_only(),
+            py::arg("memory_config")},
+
+        ttnn::pybind_overload_t{
+            [operation](const unary_backward_operation_t& self,
                const Tensor& grad_tensor,
                const ComplexTensor& input_tensor_a,
                const MemoryConfig& memory_config) -> std::vector<ComplexTensor> {
